requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow>=10.0.0
qrcode[pil]>=7.4.2
python-decouple==3.8